requirements.txt
Flask==3.0.2
python-pptx==0.6.23
Pillow==10.2.0
python-magic==0.4.27
Werkzeug==3.0.1 